+++ tasks.py
@@ -1,1 +1,4 @@
-from minchin.releaser import make_release
+try:
+    from minchin.releaser import make_release
+except ImportError:
+    print("[WARN] minchin.releaser not installed.")